回复不刷新页面登录

+++ src/main/java/lucky/yc/community/controller/PublishController.java
@@ -26,7 +26,7 @@ public class PublishController {
     private QuestionService questionService;
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name = "id") Integer id,
+    public String edit(@PathVariable(name = "id") Long id,
                        Model model) {
         QuestionDTO question = questionService.getById(id);
         model.addAttribute("title", question.getTitle());
@@ -55,7 +55,7 @@ public class PublishController {
     public String doPulish(
 //            required 是否必须
             @RequestParam(value = "title",required = false) String title, @RequestParam(value = "description",required = false) String description,
-            @RequestParam(value = "tags",required = false) String tags, @RequestParam(value = "id",required = false) Integer id, HttpServletRequest request, Model model
+            @RequestParam(value = "tags",required = false) String tags, @RequestParam(value = "id",required = false) Long id, HttpServletRequest request, Model model
     ) {
 
         model.addAttribute("title", title);
